+++ android-activity/src/game_activity/input.rs
@@ -29,6 +29,7 @@ pub enum InputEvent<'a> {
     MotionEvent(MotionEvent<'a>),
     KeyEvent(KeyEvent<'a>),
     TextEvent(crate::input::TextInputState),
+    TextAction(crate::input::TextInputAction),
 }
 
 /// A motion event.